Fix specs, add shop scopes, generic Shop

+++ db/migrate/20180922164900_drop_application_charges.rb
@@ -1,5 +1,9 @@
 class DropApplicationCharges < ActiveRecord::Migration[5.0]
   def change
+    # prevent this migration failing for apps that don't have have an
+    # application_charges table
+    return unless table_exists? :application_charges
+
     drop_table :application_charges
   end
 end
